- adds the $client property in the ClientTrait and removes it from client classes - renames the RedisTrait to MultiplesTrait - uses the MultiplesTrait in FileClient - lowers the visibility of some methods in the FileClient - updates the phpdoc
